refactor pool card
