+++ src/app/page/sideBar/communityPools/index.tsx
@@ -1,15 +1,18 @@
 import { MouseEvent, useCallback } from 'react'
 import { useDispatch, useSelector } from 'react-redux'
 import { useHistory } from 'react-router-dom'
+import LazyLoad from '@senswap/react-lazyload'
 
-import { Button, Col, Row } from 'antd'
+import { Button, Col, Empty, Row } from 'antd'
 import IonIcon from 'shared/antd/ionicon'
-import ListAllPools from './listPools'
+import PoolCard from '../components/poolCard'
 
 import configs from 'app/configs'
 import { AppDispatch, AppState } from 'app/model'
 import { handleOpenDrawer, selectPool } from 'app/model/main.controller'
 import { PoolTabs, QueryParams } from 'app/constant'
+import { useCommunityPools } from 'app/hooks/pools/useCommunityPools'
+import { useListPoolAddress } from 'app/hooks/pools/useListPoolAddress'
 
 const {
   route: { myRoute },
@@ -19,6 +22,8 @@ const CommunityPools = () => {
   const history = useHistory()
   const dispatch = useDispatch<AppDispatch>()
   const { selectedPoolAddress } = useSelector((state: AppState) => state.main)
+  const { communityPools } = useCommunityPools()
+  const { listPoolAddress } = useListPoolAddress(communityPools)
 
   const setActivePoolAddress = useCallback(
     async (address: string) => {
@@ -54,13 +59,24 @@ const CommunityPools = () => {
 
   return (
     <Row gutter={[12, 12]}>
-      <Col span={24}>
-        <ListAllPools
-          action={action}
-          selectedPoolAddress={selectedPoolAddress}
-          onClick={setActivePoolAddress}
-        />
-      </Col>
+      {!listPoolAddress.length ? (
+        <Col>
+          <Empty />
+        </Col>
+      ) : (
+        listPoolAddress.map((poolAddress) => (
+          <Col id={poolAddress} span={24} key={poolAddress}>
+            <LazyLoad height={78} overflow>
+              <PoolCard
+                poolAddress={poolAddress}
+                action={action(poolAddress)}
+                onClick={setActivePoolAddress}
+                selected={selectedPoolAddress === poolAddress}
+              />
+            </LazyLoad>
+          </Col>
+        ))
+      )}
     </Row>
   )
 }
